refactor(shorebird_cli): pull ExportOptionsPlist logic into new class
