+++ packages/shorebird_cli/bin/shorebird.dart
@@ -16,6 +16,7 @@ import 'package:shorebird_cli/src/logger.dart';
 import 'package:shorebird_cli/src/os/os.dart';
 import 'package:shorebird_cli/src/patch_diff_checker.dart';
 import 'package:shorebird_cli/src/platform.dart';
+import 'package:shorebird_cli/src/platform/platform.dart';
 import 'package:shorebird_cli/src/shorebird_artifacts.dart';
 import 'package:shorebird_cli/src/shorebird_env.dart';
 import 'package:shorebird_cli/src/shorebird_flutter.dart';
@@ -45,6 +46,7 @@ Future<void> main(List<String> args) async {
         httpClientRef,
         idevicesyslogRef,
         iosDeployRef,
+        iosRef,
         javaRef,
         loggerRef,
         osInterfaceRef,
