+++ packages/shorebird_cli/lib/src/platform/platform.dart
@@ -1,0 +1,1 @@
+export 'ios.dart';
